refactor(variable): simplify ProductionByRenewablePlant structure [7] (#3585)
