+++ src/solver/variable/include/antares/solver/variable/economy/productionByRenewablePlant.h
@@ -1,240 +1,37 @@
 // Copyright 2007-2026, RTE (https://www.rte-france.com)
 // SPDX-License-Identifier: MPL-2.0
 
-#ifndef __SOLVER_VARIABLE_ECONOMY_ProductionByRenewablePlant_H__
-#define __SOLVER_VARIABLE_ECONOMY_ProductionByRenewablePlant_H__
+#pragma once
 
-#include "antares/solver/variable/variable.h"
+#include "ByPlant_base.h"
 
 namespace Antares::Solver::Variable::Economy
 {
-struct VCardProductionByRenewablePlant
+struct ProductionByRenewablePlantTraits
 {
-    //! Caption
     static std::string Caption()
     {
         return "RES generation by plant";
     }
 
-    //! Unit
     static std::string Unit()
     {
         return "MWh";
     }
 
-    //! The short description of the variable
     static std::string Description()
     {
         return "Energy generated by all the RES clusters";
     }
 
-    //! The expecte results
-    typedef Results<R::AllYears::Average< // The average values throughout all years
-      >>
-      ResultsType;
+    static constexpr uint8_t fileLevel = Category::FileLevel::de_res;
 
-    //! The VCard to look for for calculating spatial aggregates
-    typedef VCardProductionByRenewablePlant VCardForSpatialAggregate;
-
-    //! Data Level
-    static constexpr uint8_t categoryDataLevel = Category::DataLevel::area;
-    //! File level (provided by the type of the results)
-    static constexpr uint8_t categoryFileLevel = ResultsType::categoryFile
-                                                 & (Category::FileLevel::de_res);
-    //! Precision (views)
-    static constexpr uint8_t precision = Category::all;
-    //! Indentation (GUI)
-    static constexpr uint8_t nodeDepthForGUI = +0;
-    //! Decimal precision
-    static constexpr uint8_t decimal = 0;
-    //! Number of columns used by the variable
-    static constexpr int columnCount = Category::dynamicColumns;
-    //! The Spatial aggregation
-    static constexpr uint8_t spatialAggregate = Category::spatialAggregateSum;
-    static constexpr uint8_t spatialAggregateMode = Category::spatialAggregateEachYear;
-    static constexpr uint8_t spatialAggregatePostProcessing = 0;
-    //! Intermediate values
-    static constexpr uint8_t hasIntermediateValues = 1;
-    //! Can this variable be non applicable (0 : no, 1 : yes)
-    static constexpr uint8_t isPossiblyNonApplicable = 0;
-
-    typedef IntermediateValues IntermediateValuesDeepType;
-    typedef std::vector<IntermediateValues> IntermediateValuesBaseType;
-    typedef std::vector<IntermediateValuesBaseType> IntermediateValuesType;
-
-}; // class VCard
-
-/*!
-** \brief Energy generated by all renewable clusters
-*/
-template<class NextT = Container::EndOfList>
-class ProductionByRenewablePlant: public Variable::IVariable<ProductionByRenewablePlant<NextT>,
-                                                             NextT,
-                                                             VCardProductionByRenewablePlant>
-{
-public:
-    //! Type of the next static variable
-    typedef NextT NextType;
-    //! VCard
-    typedef VCardProductionByRenewablePlant VCardType;
-    //! Ancestor
-    typedef Variable::IVariable<ProductionByRenewablePlant<NextT>, NextT, VCardType> AncestorType;
-
-    //! List of expected results
-    typedef typename VCardType::ResultsType ResultsType;
-
-    typedef VariableAccessor<ResultsType, VCardType::columnCount> VariableAccessorType;
-
-    enum
+    static size_t clusterCount(const Data::Area* area)
     {
-        //! How many items have we got
-        count = 1 + NextT::count,
-    };
-
-    template<int CDataLevel, int CFile>
-    struct Statistics
-    {
-        enum
-        {
-            count = ((VCardType::categoryDataLevel & CDataLevel
-                      && VCardType::categoryFileLevel & CFile)
-                       ? (NextType::template Statistics<CDataLevel, CFile>::count
-                          + VCardType::columnCount * ResultsType::count)
-                       : NextType::template Statistics<CDataLevel, CFile>::count),
-        };
-    };
-
-public:
-    ProductionByRenewablePlant():
-        pSize(0)
-    {
+        return area->renewable.list.enabledCount();
     }
 
-    void initializeFromStudy(Data::Study& study)
-    {
-        // Next
-        NextType::initializeFromStudy(study);
-    }
-
-    void initializeFromArea(Data::Study* study, Data::Area* area)
-    {
-        // Get the number of years in parallel
-        pNbYearsParallel = study->maxNbYearsInParallel;
-        pValuesForTheCurrentYear.resize(pNbYearsParallel);
-
-        // Get the area
-        pSize = area->renewable.list.enabledCount();
-        if (pSize)
-        {
-            AncestorType::pResults.resize(pSize);
-
-            for (unsigned int numSpace = 0; numSpace < pNbYearsParallel; numSpace++)
-            {
-                pValuesForTheCurrentYear[numSpace].resize(pSize);
-            }
-
-            for (unsigned int numSpace = 0; numSpace < pNbYearsParallel; numSpace++)
-            {
-                for (unsigned int i = 0; i != pSize; ++i)
-                {
-                    pValuesForTheCurrentYear[numSpace][i].initializeFromStudy(*study);
-                }
-            }
-
-            for (unsigned int i = 0; i != pSize; ++i)
-            {
-                AncestorType::pResults[i].initializeFromStudy(*study);
-                AncestorType::pResults[i].reset();
-            }
-        }
-        else
-        {
-            AncestorType::pResults.clear();
-        }
-        // Next
-        NextType::initializeFromArea(study, area);
-    }
-
-    size_t getMaxNumberColumns() const
-    {
-        return pSize * ResultsType::count;
-    }
-
-    void initializeFromLink(Data::Study* study, Data::AreaLink* link)
-    {
-        // Next
-        NextType::initializeFromAreaLink(study, link);
-    }
-
-    void simulationBegin()
-    {
-        // Next
-        NextType::simulationBegin();
-    }
-
-    void simulationEnd()
-    {
-        NextType::simulationEnd();
-    }
-
-    void yearBegin(unsigned int year, unsigned int numSpace)
-    {
-        // Reset the values for the current year
-        for (unsigned int i = 0; i != pSize; ++i)
-        {
-            pValuesForTheCurrentYear[numSpace][i].reset();
-        }
-        // Next variable
-        NextType::yearBegin(year, numSpace);
-    }
-
-    void yearEndBuildPrepareDataForEachThermalCluster(State& state,
-                                                      uint year,
-                                                      unsigned int numSpace)
-    {
-        // Next variable
-        NextType::yearEndBuildPrepareDataForEachThermalCluster(state, year, numSpace);
-    }
-
-    void yearEndBuild(State& state, unsigned int year, unsigned int numSpace)
-    {
-        // Next variable
-        NextType::yearEndBuild(state, year, numSpace);
-    }
-
-    void yearEnd(unsigned int year, unsigned int numSpace)
-    {
-        // Merge all results for all renewable clusters
-        {
-            for (unsigned int i = 0; i < pSize; ++i)
-            {
-                // Compute all statistics for the current year (daily,weekly,monthly)
-                pValuesForTheCurrentYear[numSpace][i].computeStatisticsForTheCurrentYear();
-            }
-        }
-        // Next variable
-        NextType::yearEnd(year, numSpace);
-    }
-
-    void computeSummary(unsigned int year, unsigned int numSpace)
-    {
-        for (unsigned int i = 0; i < pSize; ++i)
-        {
-            // Merge all those values with the global results
-            AncestorType::pResults[i].merge(year, pValuesForTheCurrentYear[numSpace][i]);
-        }
-
-        // Next variable
-        NextType::computeSummary(year, numSpace);
-    }
-
-    void hourBegin(unsigned int hourInTheYear)
-    {
-        // Next variable
-        NextType::hourBegin(hourInTheYear);
-    }
-
-    void hourForEachArea(State& state, unsigned int numSpace)
+    static void setHourlyValue(std::vector<IntermediateValues>& clusterValues, State& state)
     {
         for (const auto& renewableCluster: state.area->renewable.list.each_enabled())
         {
@@ -242,59 +39,36 @@ public:
               state.year,
               state.hourInTheYear);
 
-            pValuesForTheCurrentYear[numSpace][renewableCluster->enabledIndex]
-              .hour[state.hourInTheYear]
+            clusterValues[renewableCluster->enabledIndex].hour[state.hourInTheYear]
               += renewableClusterProduction;
         }
-
-        // Next variable
-        NextType::hourForEachArea(state, numSpace);
     }
 
-    inline void buildDigest(SurveyResults& results, int digestLevel, int dataLevel) const
+    static void buildSurveyReport(const std::vector<IntermediateValues>& clusterValues,
+                                  SurveyResults& results,
+                                  int fileLevel,
+                                  int precision)
     {
-        // Ask to build the digest to the next variable
-        NextType::buildDigest(results, digestLevel, dataLevel);
-    }
-
-    Antares::Memory::Stored<double>::ConstReturnType retrieveRawHourlyValuesForCurrentYear(
-      unsigned int column,
-      unsigned int numSpace) const
-    {
-        return pValuesForTheCurrentYear[numSpace][column].hour;
-    }
-
-    void localBuildAnnualSurveyReport(SurveyResults& results,
-                                      int fileLevel,
-                                      int precision,
-                                      unsigned int numSpace) const
-    {
-        // Initializing external pointer on current variable non applicable status
-        results.isCurrentVarNA = AncestorType::isNonApplicable;
-
-        if (AncestorType::isPrinted[0])
+        const auto& renewable = results.data.area->renewable;
+        for (const auto& cluster: renewable.list.each_enabled())
         {
-            assert(NULL != results.data.area);
-            const auto& renewable = results.data.area->renewable;
-            // Write the data for the current year
-            for (auto& cluster: renewable.list.each_enabled())
-            {
-                // Write the data for the current year
-                results.variableCaption = cluster->name();
-                results.variableUnit = VCardType::Unit();
-                pValuesForTheCurrentYear[numSpace][cluster->enabledIndex]
-                  .template buildAnnualSurveyReport<VCardType>(results, fileLevel, precision);
-            }
+            results.variableCaption = cluster->name();
+            results.variableUnit = Unit();
+            clusterValues[cluster->enabledIndex]
+              .template buildAnnualSurveyReport<VCardByPlantBase<ProductionByRenewablePlantTraits>>(
+                results,
+                fileLevel,
+                precision);
         }
     }
+};
 
-private:
-    //! Intermediate values for each year
-    typename VCardType::IntermediateValuesType pValuesForTheCurrentYear;
-    size_t pSize;
-    unsigned int pNbYearsParallel;
-}; // class ProductionByRenewablePlant
+using VCardProductionByRenewablePlant = VCardByPlantBase<ProductionByRenewablePlantTraits>;
+
+/*!
+** \brief Energy generated by all renewable clusters
+*/
+template<class NextT = Container::EndOfList>
+using ProductionByRenewablePlant = ByPlantBase<ProductionByRenewablePlantTraits, NextT>;
 
 } // namespace Antares::Solver::Variable::Economy
-
-#endif // __SOLVER_VARIABLE_ECONOMY_ProductionByRenewablePlant_H__
